A collection of small fixes.
1. Fixed path sepaerators.  Closes issue #36
2. Fixed exception emails.  Closes issue #37
3. Fixed language when installation annotations. Closes issue #23
4. Prevented user from progressing if no annotations.  Closes issue #20
5. Fixed issues where long integers were incorrectly formatted because of casting to double.  Error sometimes resulted in disparaties between the position listed on spreadsheet view, and position listed in subinspector.
6. Fixed exception that happens if one tries to remove annotations in the admin section, when no annotations exist.
7. Eliminated removal confirmation dialog in RegionDetailedListEditor if there are no regions.

+++ MedSavantClient/src/main/java/org/ut/biolab/medsavant/client/view/manage/AnnotationsPage.java
@@ -68,11 +68,11 @@ public class AnnotationsPage extends SubSectionView {
         if (view == null) {
             view = new SplitScreenView(
                     new SimpleDetailedListModel<Annotation>("Program") {
-                        @Override
-                        public Annotation[] getData() throws Exception {
-                            return ExternalAnnotationController.getInstance().getExternalAnnotations();
-                        }
-                    },
+                @Override
+                public Annotation[] getData() throws Exception {
+                    return ExternalAnnotationController.getInstance().getExternalAnnotations();
+                }
+            },
                     new ExternalAnnotationDetailedView(),
                     new ExternalAnnotationDetailedListEditor());
         }
@@ -222,16 +222,16 @@ public class AnnotationsPage extends SubSectionView {
         @Override
         public void deleteItems(List<Object[]> items) {
             try {
+                if (items != null && items.size() > 0) {
+                    Annotation an = (Annotation) items.get(0)[0];
 
-                Annotation an = (Annotation) items.get(0)[0];
+                    int response = DialogUtils.askYesNo("Confirm", "Are you sure you want to uninstall " + an.getProgram() + "?");
 
-                int response = DialogUtils.askYesNo("Confirm", "Are you sure you want to uninstall " + an.getProgram() + "?");
-
-                if (response == DialogUtils.YES) {
-                    MedSavantClient.AnnotationManagerAdapter.uninstallAnnotation(LoginController.getInstance().getSessionID(), an);
-                    DialogUtils.displayMessage("Annotation " + an.getProgram() + " uninstalled");
+                    if (response == DialogUtils.YES) {
+                        MedSavantClient.AnnotationManagerAdapter.uninstallAnnotation(LoginController.getInstance().getSessionID(), an);
+                        DialogUtils.displayMessage("Annotation " + an.getProgram() + " uninstalled");
+                    }
                 }
-
             } catch (Exception ex) {
                 ClientMiscUtils.reportError("Error uninstalling annotations", ex);
             }
